+++ src/schedule.rs
@@ -4,7 +4,7 @@
 //! and emits [`StateFlushEvent`](flush_event::StateFlushEvent).
 
 pub use apply_flush::ApplyFlushSet;
-pub use resolve_state::StateHook;
+pub use resolve_state::ResolveStateSet;
 
 pub mod apply_flush;
 pub mod detect_change;
@@ -18,7 +18,7 @@ use bevy_ecs::schedule::ScheduleLabel;
 /// The schedule that handles all [`State`](crate::state::State) flush logic, added after
 /// [`PreUpdate`](bevy_app::PreUpdate) by [`StatePlugin`](crate::setup::StatePlugin).
 ///
-/// State flush hooks run in [`StateHook<S>`] system sets and the flush is applied in
+/// State flush hooks run in [`ResolveStateSet::<S>::Flush`] and the flush is applied in
 /// [`ApplyFlushSet`].
 #[derive(ScheduleLabel, Clone, Hash, PartialEq, Eq, Debug)]
 pub struct StateFlush;